refactor info features
