+++ cypress/e2e/info/location-info.cy.ts
