+++ src/components/info/info-panel.vue
@@ -2,7 +2,7 @@
 import { useTranslation } from 'i18next-vue'
 import SidePanelLayout from '@/components/common/side-panel-layout.vue'
 import { useAppStore } from '@/stores/app.store'
-import { useInfoStore } from '@/stores/info.store'
+import { useLocationInfoStore } from '@/stores/location-info.store'
 import { storeToRefs } from 'pinia'
 import { useFeatureInfoStore } from '@/stores/feature-info.store'
 import useMap from '@/composables/map/map.composable'
@@ -12,7 +12,7 @@ import { watch, onUnmounted } from 'vue'
 
 const { t } = useTranslation()
 const appStore = useAppStore()
-const { locationInfo } = storeToRefs(useInfoStore())
+const { locationInfo } = storeToRefs(useLocationInfoStore())
 const map = useMap().olMap
 const { infoOpen } = storeToRefs(appStore)
 const { clearContent } = useFeatureInfoStore()
