+++ src/composables/info/feature-info.composable.ts
@@ -22,7 +22,7 @@ import { DrawnFeature } from '@/services/draw/drawn-feature'
 import { Pixel } from 'ol/pixel'
 import { throttle } from '@/services/utils'
 import { useMapStore } from '@/stores/map.store'
-import { useInfoStore } from '@/stores/info.store'
+import { useLocationInfoStore } from '@/stores/location-info.store'
 import { getFeatureInfoJson } from '@/services/api/api-feature-info.service'
 
 export default function useFeatureInfo() {
@@ -36,7 +36,9 @@ export default function useFeatureInfo() {
   const { drawStateActive, editStateActive, drawnFeatures } = storeToRefs(
     useDrawStore()
   )
-  const { locationInfo, isStreetviewActive } = storeToRefs(useInfoStore())
+  const { locationInfo, isStreetviewActive } = storeToRefs(
+    useLocationInfoStore()
+  )
   const { maxZoom } = storeToRefs(useMapStore())
 
   const responses = ref<FeatureInfoJSON[]>([])
