+++ src/stores/location-info.store.model.ts
